requirements.txt
# Core Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.8
bcrypt==4.1.2

# Database & ORM
psycopg2-binary==2.9.9
django-extensions==3.2.3

# Async & WebSocket Support
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Documentation
drf-spectacular==0.27.0
drf-spectacular[sidecar]==0.27.0

# Caching & Performance
django-redis==5.4.0
django-cache-machine==1.2.0

# Location & Geographic Services
geopy==2.4.1
django-location-field==2.7.3
geoip2==4.7.0

# Monitoring & Logging
sentry-sdk[django]==1.38.0
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.23.2
factory-boy==3.3.0
faker==21.0.0
coverage==7.3.2

# Rwanda Mobile Money Integration
requests==2.31.0
phonenumbers==8.13.26

# File handling & validation
Pillow==10.1.0
django-cleanup==8.0.0
django-phonenumber-field==7.3.0

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Production & Deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Health checks & monitoring
django-health-check==3.17.0

# Rwanda-specific
iso3166==2.1.1
pycountry==22.3.13

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0